// relid-IDV-Optin-OPtout/src/uniken/types/rdnaEvents.ts
export interface RDNAError {
  longErrorCode: number;
  shortErrorCode: number;
  errorString: string;
}
// ...
export interface RDNAChallengeResponse {
  status: RDNAStatus;
  session: RDNASession;
  additionalInfo: RDNAAdditionalInfo;
  challengeInfo: RDNAChallengeInfo[];
}
// ...
export interface RDNAEvent {
  challengeResponse: RDNAChallengeResponse;
  error: RDNAError;
}
// ...
export interface RDNAIDVOptInCapturedFrameConfirmationData extends RDNAEvent {
  response?: string;
  capturedImage?: string; // Base64 encoded JPEG image data
  userID?: string;
  challengeMode?: number;
}